Moves the AuthLayout to the layouts module
