--- src/components/Auth/index.js
@@ -1,7 +1,0 @@
-import Auth from './src/Auth';
-
-Auth.install = function(Vue) {
-  Vue.component(Auth.name, Auth);
-};
-
-export default Auth;
