+++ src/components/Layouts/index.js
@@ -1,5 +1,7 @@
 import AppLayout from './src/AppLayout'
+import AuthLayout from './src/AuthLayout'
 
 export {
-  AppLayout
+  AppLayout,
+  AuthLayout
 }
