+++ src/components/Layouts/src/AuthLayout/AuthLayout.vue
