+++ src/components/Layouts/src/AuthLayout/index.js
@@ -1,0 +1,7 @@
+import AuthLayout from './AuthLayout';
+
+AuthLayout.install = function(Vue) {
+  Vue.component(AuthLayout.name, AuthLayout);
+};
+
+export default AuthLayout;
