+++ tests/unit/fixtures/admin/index.js
@@ -1,5 +1,4 @@
-import { AppLayout } from '@components/Layouts'
-import AuthLayout from '@components/Auth'
+import { AppLayout, AuthLayout } from '@components/Layouts'
 import Core from '@components/Core'
 import Unauthorized from '@components/Unauthorized'
 import defaults, {
